+++ src/script/utils/string.js
@@ -47,7 +47,7 @@ export function timestampFormat(timestamp) {
 }
 
 export function dateFormat(timestamp) {
-    return timestamp.replace(/^(\d{4})(\d{2})(\d{2})$/, "$1-$2-$3");
+    return timestamp.replace(/^(\d{4})(\d{2})(\d{2})(\d{2})(\d{2})(\d{2})$/, "$1-$2-$3");
 }
 
 export function timeFormat(timestamp) {